對系統模塊進行解耦，實現了圖片的多選操作和刪除功能
